requirements.txt
openai>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
numpy>=1.24.0
pytest>=7.0.0
pytest-mock>=3.10.0
faiss-cpu>=1.7.3
